--- DAS_FRONT/.babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": ["next/babel"],
-  "plugins": [
-    [
-      "styled-components",
-      {
-        "ssr": true,
-        "displayName": true,
-        "preprocess": false
-      }
-    ]
-  ]
-}
