--- DAS_FRONT/.npmrc
@@ -1,1 +1,0 @@
-legacy-peer-deps=true
