--- DAS_FRONT/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "my-app",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next -p 3040",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "axios": "^1.2.0",
-    "next": "13.0.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-query": "^3.39.2",
-    "react-slick": "^0.29.0",
-    "recoil": "^0.7.6",
-    "slick-carousel": "^1.8.1",
-    "styled-components": "^5.3.6"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.9",
-    "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.9",
-    "@types/react-slick": "^0.23.10",
-    "@types/styled-components": "^5.1.26",
-    "babel-loader": "^9.1.0",
-    "babel-plugin-styled-components": "^2.0.7",
-    "eslint": "^8.28.0",
-    "eslint-config-next": "^13.0.5",
-    "typescript": "^4.9.3"
-  }
-}
