--- DAS_FRONT/pages/404.tsx
@@ -1,3 +1,0 @@
-export default function NotFound() {
-  return "404 error";
-}
