--- DAS_FRONT/pages/club.tsx
@@ -1,5 +1,0 @@
-const Club = () => {
-  return null;
-};
-
-export default Club;
