--- DAS_FRONT/pages/student.tsx
@@ -1,5 +1,0 @@
-const Student = () => {
-  return null;
-};
-
-export default Student;
