--- DAS_FRONT/styles/globalTheme.style.d.ts
@@ -1,9 +1,0 @@
-import "styled-components";
-
-declare module "styled-components" {
-  export interface DefaultTheme {
-    bgColor: string;
-    black: string;
-    white: string;
-  }
-}
