--- DAS_FRONT/styles/globalTheme.ts
@@ -1,7 +1,0 @@
-import { DefaultTheme } from "styled-components";
-
-export const baseTheme: DefaultTheme = {
-  bgColor: "#FAFAFA",
-  black: "#0000000",
-  white: "#ffffff",
-};
