+++ src/App.tsx
@@ -1,0 +1,5 @@
+function App() {
+  return null;
+}
+
+export default App;
