+++ src/Assets/img/custom.d.ts
