+++ src/Components/Sign/SignUpEmail.tsx
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+
+export default function SignUpEmail() {
+    return (
+        <>
+            <Container>
+                <div>asdfasdf</div>
+            </Container>
+        </>
+    )
+}
+
+const Container = styled.div`
+    height: 100vh;
+    width: 100vw;
+    background-color: #2B2B2B;
+`
