+++ src/Components/Slide.tsx
