+++ src/Components/Slider.tsx
