+++ src/Styles/Global/gloablTheme.style.ts
@@ -1,0 +1,3 @@
+import { DefaultTheme } from "styled-components";
+
+export const baseTheme: DefaultTheme = {};
