+++ src/Styles/Global/globalThemeType.style.d.ts
@@ -1,0 +1,4 @@
+import "styled-components";
+declare module "styled-components" {
+  export interface DefaultTheme {}
+}
