+++ src/index.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import App from "./App";
+import { RecoilRoot } from "recoil";
+import { ThemeProvider } from "styled-components";
+import { baseTheme } from "./Styles/Global/gloablTheme.style";
+
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
+root.render(
+  <React.StrictMode>
+    <RecoilRoot>
+      <ThemeProvider theme={baseTheme}>
+        <App />
+      </ThemeProvider>
+    </RecoilRoot>
+  </React.StrictMode>
+);
